% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_collected.R
\name{cred_collected}
\alias{cred_collected}
\title{collected}
\usage{
cred_collected(date, source = NULL, work = NULL, ...)
}
\arguments{
\item{date}{(character) a POSIX date. required.}

\item{source}{(character) alt-metric Source name. optional}

\item{work}{(character) a work DOI. optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
collected
}
\section{Pagination}{

Pagination is not done in a typical fashion with query parameters 
like \code{page = 1} and \code{per_page=10}, but rather is done 
by requesting data by date. Unfortunately, each request can only 
include data for one date.
}

\section{Sources}{

The following data sources can be passed to the \code{source} 
parameter:
\itemize{
 \item crossref_datacite
 \item facebook
 \item mendeley
 \item newsfeed
 \item reddit
 \item twitter
 \item wikipedia
 \item wordpressdotcom
 \item datacite_crossref
}
}
\examples{
\dontrun{
# all works for a single date, across all sources
res <- cred_collected(date = '2016-08-27', config=verbose())
res$`message-type`
res$`total-events`
tibble::as_data_frame(res$events)

# all works, for a single source on a single date
res <- cred_collected(date = '2016-08-27', source = 'twitter', config=verbose())
res$`message-type`
res$`total-events`
tibble::as_data_frame(res$events)

# single work, on a single date, for a single source
res <- cred_collected(date = '2016-08-27', work = '10.1107/S2056989016013359', source = 'twitter', config=verbose())
res$`message-type`
res$`total-events`
tibble::as_data_frame(res$events)
}
}

