% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_occurred.R
\name{cred_occurred}
\alias{cred_occurred}
\alias{cred_occurred_}
\title{occurred}
\usage{
cred_occurred(date, source = NULL, work = NULL, ...)

cred_occurred_(dates, source = NULL, work = NULL, ...)
}
\arguments{
\item{date}{(character) a POSIX date. required.}

\item{source}{(character) alt-metric Source name. optional}

\item{work}{(character) a work DOI. optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}

\item{dates}{(character) Two or more dates. required. dates must be of the 
form YYYY-MM-DD. You can easily create a sequence of dates like 
\code{seq(from = as.Date("2016-08-01"), to = as.Date("2016-10-18"), 
by = "day")}}
}
\description{
occurred
}
\examples{
\dontrun{
# all works for a single date, across all sources
res <- cred_occurred(date = '2016-08-27')

# all works, for a single source on a single date
res <- cred_occurred(date = '2016-08-27', source = 'twitter')
res$`message-type`
res$`total-events`
res$events

# single work, on a single date, for a single source
res <- cred_occurred(date = '2016-08-27', work = '10.1107/S2056989016013359',
  source = 'twitter')
res$`message-type`
res$`total-events`
res$events

# single work, with a range of dates, for a single source
dates <- seq(from = as.Date("2016-08-01"), to = as.Date("2016-10-18"), 
  by = "day")
res <- cred_occurred_(dates, work = '10.1056/NEJMP1608511')
res[[1]]
res[[2]]
lapply(res, "[[", "events")
do.call("rbind.data.frame", lapply(res, "[[", "events"))
}
}

