% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crevents-package.R
\docType{package}
\name{crevents-package}
\alias{crevents}
\alias{crevents-package}
\title{Client for the Crossref Events API}
\description{
Client for the Crossref Events API
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

