% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\title{Filters information}
\description{
Filters information
}
\details{
The filters:
\itemize{
\item from-occurred-date - as YYYY-MM-DD
\item until-occurred-date - as YYYY-MM-DD
\item from-collected-date - as YYYY-MM-DD
\item until-collected-date - as YYYY-MM-DD
\item subj-id - quoted URL or a DOI
\item obj-id - quoted URL or a DOI
\item subj-id.prefix - DOI prefix like 10.5555
\item obj-id.prefix - DOI prefix like 10.5555
\item subj-id.domain - domain of the subj_id e.g. en.wikipedia.org
\item obj-id.domain - domain of the obj_url e.g. en.wikipedia.org
\item subj.url - quoted full URL
\item obj.url - quoted full URL
\item subj.url.domain - domain of the optional subj.url, if present
e.g. en.wikipedia.org
\item obj.url.domain - domain of the optional obj.url, if present
e.g. en.wikipedia.org
\item subj.alternative-id - optional subj.alternative-id
\item obj.alternative-id - optional obj.alternative-id
\item relation - relation type ID
\item source - source ID
}
}
