% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cred_evidence.R
\name{cred_evidence}
\alias{cred_evidence}
\title{collected}
\usage{
cred_evidence(id, ...)
}
\arguments{
\item{id}{(character) an event id. required.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
collected
}
\section{Collected vs. Occurred}{

\itemize{
 \item \strong{collected} is useful when you want to run a daily query to 
 fetch all Events for a given filter and you want to be sure you always 
 have all available Events
 \item \strong{collected} is useful when you want to reference a dataset 
 and be sure it never changes
 \item \strong{occurred} is useful when you want to retrieve Events that
 occurred at a particular time
 \item when using \strong{occurred} you should be aware that new Events 
 may be collected at any time in the future that occurred at a date 
 in the past
}
}

\section{Pagination}{

Pagination is not done in a typical fashion with query parameters 
like \code{page = 1} and \code{per_page=10}, but rather is done 
by requesting data by date. Unfortunately, each request can only 
include data for one date.
}

\section{Sources}{

The following data sources can be passed to the \code{source} 
parameter:
\itemize{
 \item crossref_datacite
 \item facebook
 \item mendeley
 \item newsfeed
 \item reddit
 \item twitter
 \item wikipedia
 \item wordpressdotcom
 \item datacite_crossref
}
}
\examples{
\dontrun{
# get some data
res <- cred_collected(date = '2016-08-27', source = 'twitter', config=verbose())

# get a single event id
cred_evidence(id = "06630d1f-3add-4478-a2c8-faa38728e0d8", config=verbose())

# get many events
foo <- lapply(res$events$id[1:5], cred_evidence)
}
}

