% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_occurred.R
\name{cred_occurred}
\alias{cred_occurred}
\title{occurred}
\usage{
cred_occurred(date, source = NULL, work = NULL, ...)
}
\arguments{
\item{date}{(character) a POSIX date. required.}

\item{source}{(character) alt-metric Source name. optional}

\item{work}{(character) a work DOI. optional}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
occurred
}
\examples{
\dontrun{
# all works for a single date, across all sources
res <- cred_occurred(date = '2016-08-27', config=verbose())

# all works, for a single source on a single date
res <- cred_occurred(date = '2016-08-27', source = 'twitter', config=verbose())

# single work, on a single date, for a single source
res <- cred_occurred(date = '2016-08-27', work = '10.1107/S2056989016013359', source = 'twitter', config=verbose())
res$`message-type`
res$`total-events`
res$events
}
}

