% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crevents-package.R
\docType{package}
\name{crevents-package}
\alias{crevents-package}
\alias{crevents}
\title{Client for the Crossref Events API}
\description{
Documentation for the Crossref Events Data service at
\url{https://www.eventdata.crossref.org/guide/index.html} and
\url{https://www.eventdata.crossref.org/guide/}
}
\author{
Scott Chamberlain
}
\keyword{package}
